backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
gunicorn==23.0.0
python-dotenv==1.0.1
google-cloud-speech==2.28.0
google-generativeai==0.8.3
elevenlabs==1.12.1
pymongo==4.10.1
motor==3.6.0
firebase-admin==6.6.0
textblob==0.18.0.post0
pydantic-settings==2.6.1
